+++ Eradication/GenomeMarkers.h
@@ -31,7 +31,7 @@ namespace Kernel
         virtual int Initialize( const std::vector<std::string>& rAllGenomeMarkerNames ) override;
 
         // Return an integer that has the bits set for each marker name given
-        virtual uint32_t CreateBits( const std::vector<std::string>& rGenomeMarkerNames ) const override;
+        virtual uint64_t CreateBits( const std::vector<std::string>& rGenomeMarkerNames ) const override;
 
         // Return all the marker names but in a set - used with initConfigTypeMap() when an object is reading
         // a subset of marker names.
@@ -42,25 +42,25 @@ namespace Kernel
         // four combinations: NoMarkers, A, B, C, A-B. A-C, B-C, A-B-C.  Notice that the string representing
         // the markers has dashes between the marker names.  Also notice that "NoMarkers" is used for the empty set.
         // This method was initially created for reports that needed statistics on all of the combinations.
-        virtual std::vector<std::pair<std::string, uint32_t>> CreatePossibleCombinations() const override;
+        virtual std::vector<std::pair<std::string, uint64_t>> CreatePossibleCombinations() const override;
 
         // Return the number of markers
-        virtual uint32_t Size() const override;
+        virtual uint64_t Size() const override;
 
         // Return the bits for the given name.  This is an integer with the bit set for this marker.
-        virtual uint32_t GetBits( const std::string& rName ) const override;
+        virtual uint64_t GetBits( const std::string& rName ) const override;
 
         // Return all of the combinatinos (order-independent) of given list of names.
         // This does not include the empty set.
         static std::vector<std::vector<std::string>> GetCombinations( const std::vector<std::string>& rNames );
 
         // Combine the bits of the two genetic ID's into one new one.
-        static uint32_t FromOutcrossing( RANDOMBASE* pRNG, uint32_t id1, uint32_t id2 );
+        static uint64_t FromOutcrossing( RANDOMBASE* pRNG, uint64_t id1, uint64_t id2 );
 
     protected:
-        const std::pair<std::string, uint32_t>* Get( const std::string& rName ) const;
+        const std::pair<std::string, uint64_t>* Get( const std::string& rName ) const;
 
-        std::map<std::string,std::pair<std::string,uint32_t>> m_MarkerMap;  // this maps lowercase name to input name/bits
+        std::map<std::string,std::pair<std::string,uint64_t>> m_MarkerMap;  // this maps lowercase name to input name/bits
         std::set<std::string> m_NameSet;
     };
 }
