+++ Eradication/IGenomeMarkers.h
@@ -25,7 +25,7 @@ namespace Kernel
         virtual int Initialize( const std::vector<std::string>& rAllGenomeMarkerNames ) = 0;
 
         // Return an integer that has the bits set for each marker name given
-        virtual uint32_t CreateBits( const std::vector<std::string>& rGenomeMarkerNames ) const = 0;
+        virtual uint64_t CreateBits( const std::vector<std::string>& rGenomeMarkerNames ) const = 0;
 
         // Return all the marker names but in a set - used with initConfigTypeMap() when an object is reading
         // a subset of marker names.
@@ -36,12 +36,12 @@ namespace Kernel
         // four combinations: NoMarkers, A, B, C, A-B. A-C, B-C, A-B-C.  Notice that the string representing
         // the markers has dashes between the marker names.  Also notice that "NoMarkers" is used for the empty set.
         // This method was initially created for reports that needed statistics on all of the combinations.
-        virtual std::vector<std::pair<std::string, uint32_t>> CreatePossibleCombinations() const = 0;
+        virtual std::vector<std::pair<std::string, uint64_t>> CreatePossibleCombinations() const = 0;
 
         // Return the number of markers
-        virtual uint32_t Size() const = 0;
+        virtual uint64_t Size() const = 0;
 
         // Return the bits for the given name.  This is an integer with the bit set for this marker.
-        virtual uint32_t GetBits( const std::string& rName ) const = 0;
+        virtual uint64_t GetBits( const std::string& rName ) const = 0;
     };
 }
